// generated/src/app/api-models/profsvc-claim-header-edi.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class ProfsvcClaimHeaderEdi {
  transactionSetId: string;
  seqClaimId: number;
  claimNumber: string;
  primarySvcDate: Date;
  claimThruDate: Date;
  authNumber: number;
  secondaryAuth: string;
  seqMembId: number;
  memberAge: number;
  memberGender: string;
  seqPcpId: number;
  pcpType: string;
  pcpSpec: string;
  pcpIpaId: string;
  seqGroupId: number;
  planCode: string;
  lineOfBusiness: string;
  panelId: string;
  riderString: string;
  eligibleThruDate: Date;
  eligStatus: string;
  seqProvId: number;
  seqProvAddress: number;
  providerParStat: string;
  providerType: string;
  providerSpec: string;
  providerIpaId: string;
  providerPcpFlag: string;
  seqRefProvId: number;
  refProvType: string;
  refProvSpec: string;
  refProvParStat: string;
  refProvIpaId: string;
  seqVendId: number;
  seqVendAddress: number;
  totalBilledAmt: number;
  placeOfService1: string;
  placeOfService2: string;
  placeOfService3: string;
  serviceReason1: string;
  diagnosis1: string;
  diagnosis2: string;
  diagnosis3: string;
  diagnosis4: string;
  dateReceived: Date;
  userDefined1: string;
  userDefined2: string;
  batchNumber: string;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
  seqProvContract: number;
  paySubscriber: string;
  memberLastName: string;
  memberFirstName: string;
  memberDateOfBirth: Date;
  memberSocialSecNo: string;
  subscriberId: string;
  personNumber: string;
  extProvId: string;
  extProvIdType: string;
  extVendId: string;
  extVendIdType: string;
}